mismatch_check/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10
opencv-python>=4.8.0
